requirements.txt
crewai
langchain
pymupdf
chromadb
groq
python-dotenv
fastapi
uvicorn
python-multipart
python-docx
